Move type-system analyses into a subdirectory
Changes:
- Moved type-system related analyses into a subdirectory "typesystem"
- Split type constraints declaration from the component that generates
them. The constraints are now declared in `TypeConstraints` and they
are generated/discovered in `TypeConstrainsAnalysis`
