+++ src/CMakeLists.txt
@@ -54,7 +54,6 @@ set(SOUFFLE_SOURCES
     ast/analysis/Functor.cpp
     ast/analysis/Ground.cpp
     ast/analysis/IOType.cpp
-    ast/analysis/PolymorphicObjects.cpp
     ast/analysis/PrecedenceGraph.cpp
     ast/analysis/ProfileUse.cpp
     ast/analysis/RecursiveClauses.cpp
@@ -62,12 +61,15 @@ set(SOUFFLE_SOURCES
     ast/analysis/RelationDetailCache.cpp
     ast/analysis/RelationSchedule.cpp
     ast/analysis/SCCGraph.cpp
-    ast/analysis/SumTypeBranches.cpp
     ast/analysis/TopologicallySortedSCCGraph.cpp
-    ast/analysis/Type.cpp
-    ast/analysis/TypeConstraints.cpp
-    ast/analysis/TypeSystem.cpp
-    ast/analysis/TypeEnvironment.cpp
+    ast/analysis/typesystem/PolymorphicObjects.cpp
+    ast/analysis/typesystem/SumTypeBranches.cpp
+    ast/analysis/typesystem/Type.cpp
+    ast/analysis/typesystem/TypeConstraints.cpp
+    ast/analysis/typesystem/TypeConstrainsAnalysis.cpp
+    ast/analysis/typesystem/TypeConstrainsAnalysis.h
+    ast/analysis/typesystem/TypeSystem.cpp
+    ast/analysis/typesystem/TypeEnvironment.cpp
     ast/transform/AddNullariesToAtomlessAggregates.cpp
     ast/transform/ComponentChecker.cpp
     ast/transform/ComponentInstantiation.cpp
