+++ src/ast/analysis/typesystem/SumTypeBranches.cpp
@@ -14,14 +14,14 @@
  *
  ***********************************************************************/
 
-#include "ast/analysis/SumTypeBranches.h"
+#include "ast/analysis/typesystem/SumTypeBranches.h"
 #include "ast/AlgebraicDataType.h"
 #include "ast/BranchType.h"
 #include "ast/Program.h"
 #include "ast/TranslationUnit.h"
 #include "ast/Type.h"
-#include "ast/analysis/TypeEnvironment.h"
-#include "ast/analysis/TypeSystem.h"
+#include "ast/analysis/typesystem/TypeEnvironment.h"
+#include "ast/analysis/typesystem/TypeSystem.h"
 #include "ast/utility/Visitor.h"
 #include <vector>
