+++ src/ast/analysis/typesystem/TypeConstrainsAnalysis.h
@@ -8,69 +8,15 @@
 
 /************************************************************************
  *
- * @file TypeConstraints.h
+ * @file TypeConstraintsAnalysis.h
  *
  ***********************************************************************/
 
 #pragma once
 
-#include "ast/analysis/Constraint.h"
-#include "ast/analysis/ConstraintSystem.h"
-#include "ast/analysis/SumTypeBranches.h"
-#include "ast/analysis/Type.h"
-#include "ast/analysis/TypeEnvironment.h"
-#include "ast/analysis/TypeSystem.h"
-#include "ast/utility/Utils.h"
-#include "souffle/utility/StringUtil.h"
+#include "ast/analysis/typesystem/TypeConstraints.h"
 
 namespace souffle::ast::analysis {
-
-// -----------------------------------------------------------------------------
-//                          Type Deduction Lattice
-// -----------------------------------------------------------------------------
-
-/**
- * An implementation of a meet operation between sets of types computing
- * the set of pair-wise greatest common subtypes.
- */
-struct sub_type {
-    bool operator()(TypeSet& a, const TypeSet& b) const {
-        // compute result set
-        TypeSet greatestCommonSubtypes = getGreatestCommonSubtypes(a, b);
-
-        // check whether a should change
-        if (greatestCommonSubtypes == a) {
-            return false;
-        }
-
-        // update a
-        a = greatestCommonSubtypes;
-        return true;
-    }
-};
-
-/**
- * A factory for computing sets of types covering all potential types.
- */
-struct all_type_factory {
-    TypeSet operator()() const {
-        return TypeSet(true);
-    }
-};
-
-/**
- * The type lattice forming the property space for the Type analysis. The
- * value set is given by sets of types and the meet operator is based on the
- * pair-wise computation of greatest common subtypes. Correspondingly, the
- * bottom element has to be the set of all types.
- */
-struct type_lattice : public property_space<TypeSet, sub_type, all_type_factory> {};
-
-/** The definition of the type of variable to be utilized in the type analysis */
-using TypeVar = ConstraintAnalysisVar<type_lattice>;
-
-/** The definition of the type of constraint to be utilized in the type analysis */
-using TypeConstraint = std::shared_ptr<Constraint<TypeVar>>;
 
 /**
  * Constraint analysis framework for types.
