+++ src/ast/transform/ExpandEqrels.cpp
@@ -38,7 +38,7 @@
 #include "ast/UnnamedVariable.h"
 #include "ast/UserDefinedFunctor.h"
 #include "ast/Variable.h"
-#include "ast/analysis/PolymorphicObjects.h"
+#include "ast/analysis/typesystem/PolymorphicObjects.h"
 #include "ast/utility/Utils.h"
 #include "ast/utility/Visitor.h"
 #include "souffle/BinaryConstraintOps.h"
