+++ src/ast/transform/IOAttributes.h
@@ -26,8 +26,8 @@
 #include "ast/Relation.h"
 #include "ast/TranslationUnit.h"
 #include "ast/Type.h"
-#include "ast/analysis/TypeEnvironment.h"
-#include "ast/analysis/TypeSystem.h"
+#include "ast/analysis/typesystem/TypeEnvironment.h"
+#include "ast/analysis/typesystem/TypeSystem.h"
 #include "ast/transform/Transformer.h"
 #include "ast/utility/Utils.h"
 #include "ast/utility/Visitor.h"
