+++ src/ast/transform/MagicSet.cpp
@@ -32,10 +32,10 @@
 #include "ast/TranslationUnit.h"
 #include "ast/UnnamedVariable.h"
 #include "ast/analysis/IOType.h"
-#include "ast/analysis/PolymorphicObjects.h"
 #include "ast/analysis/PrecedenceGraph.h"
 #include "ast/analysis/RelationDetailCache.h"
 #include "ast/analysis/SCCGraph.h"
+#include "ast/analysis/typesystem/PolymorphicObjects.h"
 #include "ast/utility/BindingStore.h"
 #include "ast/utility/Utils.h"
 #include "ast/utility/Visitor.h"
