+++ src/ast/transform/MaterializeAggregationQueries.cpp
@@ -29,8 +29,8 @@
 #include "ast/Variable.h"
 #include "ast/analysis/Aggregate.h"
 #include "ast/analysis/Ground.h"
-#include "ast/analysis/Type.h"
-#include "ast/analysis/TypeSystem.h"
+#include "ast/analysis/typesystem/Type.h"
+#include "ast/analysis/typesystem/TypeSystem.h"
 #include "ast/utility/Utils.h"
 #include "ast/utility/Visitor.h"
 #include "souffle/TypeAttribute.h"
