+++ src/ast/transform/SemanticChecker.cpp
@@ -60,9 +60,9 @@
 #include "ast/analysis/PrecedenceGraph.h"
 #include "ast/analysis/RecursiveClauses.h"
 #include "ast/analysis/SCCGraph.h"
-#include "ast/analysis/Type.h"
-#include "ast/analysis/TypeEnvironment.h"
-#include "ast/analysis/TypeSystem.h"
+#include "ast/analysis/typesystem/Type.h"
+#include "ast/analysis/typesystem/TypeEnvironment.h"
+#include "ast/analysis/typesystem/TypeSystem.h"
 #include "ast/transform/GroundedTermsChecker.h"
 #include "ast/transform/TypeChecker.h"
 #include "ast/utility/Utils.h"
